Fix exception when running `cvd start --help`
`cvd start` looks for `ANDROID_HOST_OUT` in a map of environment
variables, expecting it will be there.

WIP for a better solution going forward.
The current version fails in an unexpected way with a terrible error
message if run from a non-`HOME` directory, with `ANDROID_HOST_OUT`
unset, but the host tools are present.  It finds `cvd_internal_start` to
run help, but `cvd_internal_start` cannot find `assemble_cvd`.