Split out blocking logic into a helper
To make the flow of `GetBuild` simpler for the sake of upcoming
refactors.

Removed extra `LOG(DEBUG)` messages to avoid log spamming during
non-blocking execution.  When the blocking logic triggers the status
is still logged.

Duplicated the `BuildStatus` HTTP request because that method will be
updated and diverge in a following commit.

Bug:440398826