Add library to gather host information for metrics
Some duplication of `host/commands/cvd/metrics/utils.h` and
`common/libs/utils/architecture.h` due to both locations not allowing a
clean failure as a `Result`.

All of the new metrics code needs to be able to fail in a way that can
be caught to avoid program failure AND returns information we can review
and act on.  `CHECK` calls could cause failures users should not need to
deal with, and default values hide failures or gaps in data gathering.

Bug: 441150467