WIP move metrics transmission logic to new binary
The idea is to move all of the Google-specific logic into this new
binary, then make the binary part of a new package that enables metrics
transmission when installed (and the binary is present).

Still need to delete `host/libs/metrics/metrics_transmitter.h/cc`, and
clean up to fix the new boundary of `metrics_conversion.h/cc` and
`metrics_writer.h/cc` to be our `CuttlefishLogEvent` proto instead of
the Google protos.

I also think the former `metrics_transmitter.h/cc` code should go in a
tiny library that is imported here into the binary's `main` logic.
The extracted `metrics_conversion.cc` code, as well, to keep this new
file focused on the logging/flags processing/high-level transmission.

Bug: 450053679